update: use vlm parser to extract keywords
chore: pass converter

chore: drop papers

+++ util/process_paper.py
@@ -21,6 +21,10 @@ from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.common.action_chains import ActionChains
 from selenium.webdriver.common.by import By
+from docling.datamodel import vlm_model_specs
+from docling.datamodel.base_models import InputFormat
+from docling.document_converter import DocumentConverter, PdfFormatOption
+from docling.pipeline.vlm_pipeline import VlmPipeline
 
 from util.const import PARSED_DIR, PDF_DIR, PDF_LINK_CSV, TITLE_LIST
 
@@ -74,6 +78,7 @@ def download_pdf(path: Path, url: str) -> None:
     except Exception as e:
         logger.info(f"Failed download {url}: {e}")
 
+
 def extract_sections_with_content(parsed_data: dict) -> dict:
     sections = {}
     current_section = None
@@ -101,15 +106,37 @@ def extract_sections_with_content(parsed_data: dict) -> dict:
 
     return sections
 
-def parse_pdf(path: Path, hashed: str) -> None:
-    converter = DocumentConverter()
+def parse_pdf(converter: DocumentConverter, hashed: str, use_cache=True) -> None:
+    if use_cache and (PARSED_DIR / f"{hashed}.json").exists():
+        return
+
     try:
         parsed = converter.convert(PDF_DIR / f"{hashed}.pdf").document
         section_contents = extract_sections_with_content(parsed.export_to_dict())
         with open(PARSED_DIR / f"{hashed}.json", 'w', encoding='utf-8') as f:
             json.dump(section_contents, f)
     except Exception as e:
-        logger.info(f"Parse error {path}: {e}")
+        logger.info(f"Parse error {hashed}: {e}")
+
+def extract_keywords(converter: DocumentConverter, hashed: str) -> None:
+    try:
+        parsed = converter.convert(source=PDF_DIR / f"{hashed}.pdf", page_range=(1, 1)).document
+        section_contents = extract_sections_with_content(parsed.export_to_dict())
+        keyword_key = [key for key in section_contents.keys() if 'keyword' in key.lower()]
+
+        if not keyword_key:
+            logger.warning(f"No keywords found in {hashed}")
+            return
+
+        with open(PARSED_DIR / f"{hashed}.json", 'r+', encoding='utf-8') as f:
+            data = json.load(f)
+            data['keywords_parsed'] = [item.strip() for item in section_contents[keyword_key[0]].split(',')]
+            f.seek(0)
+            f.truncate()
+            json.dump(data, f, ensure_ascii=False, indent=2)
+
+    except Exception as e:
+        logger.info(f"Parse error {hashed}: {e}")
 
 
 # ==== Selenium Helpers ====
@@ -251,6 +278,7 @@ if __name__ == '__main__':
     parser.add_argument('--fetch_links', action='store_true')
     parser.add_argument('--download_pdfs', action='store_true')
     parser.add_argument('--parse_pdfs', action='store_true')
+    parser.add_argument('--extract_keywords', action='store_true')
     args = parser.parse_args()
 
     if args.fetch_titles:
@@ -269,5 +297,21 @@ if __name__ == '__main__':
     if args.parse_pdfs:
         titles = fetch_acm_titles(max_pages=50, use_cache=True)
         df_links = fetch_arxiv_links(titles, use_cache=True)
+        converter = DocumentConverter()
         for _, row in tqdm.tqdm(df_links.dropna(subset=['PDF_Link']).iterrows(), total=len(df_links)):
-            parse_pdf(PARSED_DIR / f"{row['Hashed']}.pdf", row['Hashed'])
+            parse_pdf(converter, row['Hashed'], use_cache=False)
+
+    if args.extract_keywords:
+        titles = fetch_acm_titles(max_pages=50, use_cache=True)
+        df_links = fetch_arxiv_links(titles, use_cache=True)
+
+        vlm_converter = DocumentConverter(
+            format_options={
+                InputFormat.PDF: PdfFormatOption(
+                    pipeline_cls=VlmPipeline,
+                ),
+            }
+        )
+        for _, row in tqdm.tqdm(df_links.dropna(subset=['PDF_Link']).iterrows(), total=len(df_links)):
+            parse_pdf(converter, row['Hashed'], use_cache=True)
+            extract_keywords(vlm_converter, row['Hashed'])
